requirements.txt
requests
discord.py
python-decouple
tabulate
networkx
matplotlib
